Pull all Google Admob to a single loader
