+++ mobile/src/main/java/com/murati/audiobook/ui/AboutActivity.java
@@ -26,18 +26,15 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdView;
-import com.google.android.gms.ads.MobileAds;
 import com.murati.audiobook.BuildConfig;
 import com.murati.audiobook.R;
 
+import com.murati.audiobook.utils.AdHelper;
 import com.murati.audiobook.utils.FeedbackHelper;
 import com.murati.audiobook.utils.LogHelper;
 
-import com.huawei.hms.ads.AdParam;
 import com.huawei.hms.ads.banner.BannerView;
-import com.huawei.hms.ads.AdListener;
 
 import com.murati.audiobook.utils.MobileServicesHelper;
 
@@ -82,63 +79,17 @@ public class AboutActivity extends BaseActivity {
             public void onClick(View v) { openBrowser("https://www.patreon.com/murati");}
         });
 
-        // TODO: Ad loader
-        // Huawei load as a fragment
+        // Todo: Huawei load as a fragment
         try {
-            if(true || MobileServicesHelper.isHmsAvailable(this)) {
-                huaweiAdView = findViewById(R.id.huaweiAdView);
-                huaweiAdView.setVisibility(View.VISIBLE);
-                // Create an ad request to load an ad.
-                AdParam adParam = new AdParam.Builder().build();
-                huaweiAdView.loadAd(adParam);
-                huaweiAdView.setAdListener(new AdListener() {
-                    @Override
-                    public void onAdLoaded() {
-                        // Called when an ad is loaded successfully.
-                        Log.d(TAG, "onAdLoaded");
-                    }
-
-                    @Override
-                    public void onAdFailed(int errorCode) {
-                        // Called when an ad fails to be loaded.
-                        Log.d(TAG, "onAdFailed");
-                    }
-
-                    @Override
-                    public void onAdOpened() {
-                        // Called when an ad is opened.
-                        Log.d(TAG, "onAdOpened");
-                    }
-
-                    @Override
-                    public void onAdClicked() {
-                        // Called when a user taps an ad.
-                        Log.d(TAG, "onAdClicked");
-                    }
-
-                    @Override
-                    public void onAdLeave() {
-                        // Called when a user has left the app.
-                        Log.d(TAG, "onAdLeave");
-                    }
-
-                    @Override
-                    public void onAdClosed() {
-                        // Called when an ad is closed.
-                        Log.d(TAG, "onAdClosed");
-                    }
-                });
-            }
+            AdHelper.loadHuaweiAdkitToView(this, R.id.huaweiAdView);
+            // MobileServicesHelper.isHmsAvailable(this)
         } catch (Exception ex) {
             Log.e(TAG, ex.getMessage());
         }
 
         // Google
         try {
-            MobileAds.initialize(this, BuildConfig.ADMOB_APP_ID);
-            mAdView = findViewById(R.id.adView);
-            AdRequest adRequest = new AdRequest.Builder().build();
-            mAdView.loadAd(adRequest);
+            AdHelper.loadGoogleAdmodToView(this, R.id.adView);
         } catch (Exception ex) {
             Log.e(TAG, ex.getMessage());
         }
