+++ mobile/src/main/java/com/murati/audiobook/ui/FullScreenPlayerActivity.java
@@ -50,6 +50,7 @@ import com.murati.audiobook.BuildConfig;
 import com.murati.audiobook.MusicService;
 import com.murati.audiobook.OfflineBookService;
 import com.murati.audiobook.R;
+import com.murati.audiobook.utils.AdHelper;
 import com.murati.audiobook.utils.BitmapHelper;
 import com.murati.audiobook.utils.LogHelper;
 import com.murati.audiobook.utils.PlaybackHelper;
@@ -158,14 +159,7 @@ public class FullScreenPlayerActivity extends ActionBarCastActivity {
         initializeToolbar();
 
         try {
-            MobileAds.initialize(this, BuildConfig.ADMOB_APP_ID);
-            mAdView = findViewById(R.id.adView);
-            //if (!BuildConfig.DEBUG) {
-            //mAdView.setAdSize(AdSize.BANNER);
-            //mAdView.setAdUnitId(getString(R.string.admob_unit_id_1));
-            AdRequest adRequest = new AdRequest.Builder().build();
-            mAdView.loadAd(adRequest);
-            //}
+            AdHelper.loadGoogleAdmodToView(this, R.id.adView);
         } catch (Exception ex) {
             Log.e(TAG, ex.getMessage());
         }
