+++ mobile/src/main/res/layout/activity_player.xml
@@ -78,7 +78,7 @@
                     tools:layout="@layout/fragment_playback_controls" />
 
                 <com.google.android.gms.ads.AdView
-                    android:id="@+id/control_ad"
+                    android:id="@+id/adView"
                     android:visibility="gone"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
