+++ mobile/src/main/res/layout/fragment_list_item.xml
@@ -13,6 +13,8 @@
   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   See the License for the specific language governing permissions and
   limitations under the License.
+
+  TODO: cleanup
   -->
 
 <androidx.cardview.widget.CardView
